requirements-dev.txt
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
Pillow==6.1.0
pkg-resources==0.0.0
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0